firmware: move max5864 driver instance to max5864.x
