+++ firmware/common/max5864.c
@@ -24,6 +24,35 @@
 
 #include <stdint.h>
 
+#include <libopencm3/lpc43xx/ssp.h>
+
+#include "max5864_target.h"
+#include "spi_bus.h"
+
+/* Driver instance. */
+ssp_config_t ssp_config_max5864 = {
+	/* FIXME speed up once everything is working reliably */
+	/*
+	// Freq About 0.0498MHz / 49.8KHz => Freq = PCLK / (CPSDVSR * [SCR+1]) with PCLK=PLL1=204MHz
+	const uint8_t serial_clock_rate = 32;
+	const uint8_t clock_prescale_rate = 128;
+	*/
+	// Freq About 4.857MHz => Freq = PCLK / (CPSDVSR * [SCR+1]) with PCLK=PLL1=204MHz
+	.data_bits = SSP_DATA_8BITS,
+	.serial_clock_rate = 21,
+	.clock_prescale_rate = 2,
+};
+
+max5864_driver_t max5864 = {
+	.bus = &spi_bus_ssp1,
+	.target_init = max5864_target_init,
+};
+
+void ssp1_set_mode_max5864(void)
+{
+	spi_bus_start(max5864.bus, &ssp_config_max5864);
+}
+
 static void max5864_write(max5864_driver_t* const drv, uint8_t value)
 {
 	spi_bus_transfer(drv->bus, &value, 1);
