+++ firmware/common/max5864.h
@@ -23,6 +23,7 @@
 #pragma once
 
 #include "spi_bus.h"
+#include "spi_ssp.h"
 
 typedef struct _max5864_driver_t {
 	spi_bus_t* const bus;
@@ -37,3 +38,8 @@ void max5864_idle(max5864_driver_t* const drv);
 void max5864_rx(max5864_driver_t* const drv);
 void max5864_tx(max5864_driver_t* const drv);
 void max5864_xcvr(max5864_driver_t* const drv);
+
+/* Driver instance. */
+extern ssp_config_t ssp_config_max5864;
+extern max5864_driver_t max5864;
+void ssp1_set_mode_max5864(void);
